making a better backend business logic : \n conserned entities : candidat, car, ecole, entraineur and lesson. \n comment : true, unicity-validation : true, jdl-updated : true, frontend-updated : false, liquibase : false \n author : Taleb Mohammed Housseyn 24-03-19 01:53 maghnia

+++ src/main/java/dz/talcorp/ae/web/rest/CandidatResource.java
@@ -1,6 +1,7 @@
 package dz.talcorp.ae.web.rest;
 import dz.talcorp.ae.service.CandidatService;
 import dz.talcorp.ae.web.rest.errors.BadRequestAlertException;
+import dz.talcorp.ae.web.rest.errors.CustomParameterizedException;
 import dz.talcorp.ae.web.rest.util.HeaderUtil;
 import dz.talcorp.ae.web.rest.util.PaginationUtil;
 import dz.talcorp.ae.service.dto.CandidatDTO;
@@ -10,7 +11,6 @@ import org.slf4j.LoggerFactory;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -51,6 +51,18 @@ public class CandidatResource {
         if (candidatDTO.getId() != null) {
             throw new BadRequestAlertException("A new candidat cannot already have an ID", ENTITY_NAME, "idexists");
         }
+        /**
+         * checking unique nid
+         */
+        if(candidatService.checkNID(candidatDTO.getNid())){
+            throw new CustomParameterizedException("candidat.EK_CA_01", candidatDTO.getNid());
+        }
+        /**
+         * Check candidat is 17 years old or more
+         */
+        if(candidatService.checkCandidatMature(candidatDTO.getDateNaissance())){
+            throw new BadRequestAlertException("a new candidat must be mature 17 years old or more", ENTITY_NAME, "candidat.EK_CA_02");
+        }
         CandidatDTO result = candidatService.save(candidatDTO);
         return ResponseEntity.created(new URI("/api/candidats/" + result.getId()))
             .headers(HeaderUtil.createEntityCreationAlert(ENTITY_NAME, result.getId().toString()))
@@ -71,6 +83,18 @@ public class CandidatResource {
         log.debug("REST request to update Candidat : {}", candidatDTO);
         if (candidatDTO.getId() == null) {
             throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idnull");
+        }
+        /**
+         * checking unique nid
+         */
+        if(candidatService.checkNID(candidatDTO.getNid())){
+            throw new CustomParameterizedException("candidat.EK_CA_01", candidatDTO.getNid());
+        }
+        /**
+         * Check candidat is 17 years old or more
+         */
+        if(candidatService.checkCandidatMature(candidatDTO.getDateNaissance())){
+            throw new BadRequestAlertException("a new candidat must be mature 17 years old or more", ENTITY_NAME, "candidat.EK_CA_02");
         }
         CandidatDTO result = candidatService.save(candidatDTO);
         return ResponseEntity.ok()
@@ -107,13 +131,29 @@ public class CandidatResource {
 
     /**
      * DELETE  /candidats/:id : delete the "id" candidat.
+     * 
+     * this need to be updated
+     * 
+     * a failed delete should have some reporting to failure cause
+     * 
+     * 1- relation with a lesson
+     * 2- relation with an exam
      *
+     * this should not be implemented in the service delete cuz maybe in futur I will add force delete option 
+     * and to avoid any conflict with jhipster merging code when using git
+     * 
      * @param id the id of the candidatDTO to delete
      * @return the ResponseEntity with status 200 (OK)
      */
     @DeleteMapping("/candidats/{id}")
     public ResponseEntity<Void> deleteCandidat(@PathVariable Long id) {
         log.debug("REST request to delete Candidat : {}", id);
+        if (candidatService.checkLessonRelation(id)) {
+            throw new BadRequestAlertException("delete impossible of candidat due to relation with lesson entity", ENTITY_NAME, "candidat.EK_DLR_01");
+        }
+        if (candidatService.checkExamRelation(id)) {
+            throw new BadRequestAlertException("delete impossible of candidat due to relation with examenInfo entity", ENTITY_NAME, "candidat.EK_DER_02");
+        }
         candidatService.delete(id);
         return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert(ENTITY_NAME, id.toString())).build();
     }
